Binning dialog is functional
The new BinningDialog object will create a new particle stack from an old one, with the user being able to specify their desired pixel size, with calculations performed to verify that the ratio is accurate based on the box size, which is the limiting factor.

This commit still contains many comments, as there is logic that will be useful for addition of features that will be related to this one.